requirements.txt
flask
line-bot-sdk
langchain
openai
tiktoken
chromadb
sentence-transformers
requests
python-dotenv
feedparser
